requirements.txt
pandas>=2.1.0
numpy>=1.26.0
python-dateutil>=2.8.0
requests>=2.31.0
yfinance>=0.2.37
schedule>=1.2.0
pytz>=2024.1
# Phase 4
anthropic>=0.50.0
streamlit>=1.35.0
altair>=5.3.0
python-dotenv>=1.0.0
